Modularize dev_rlwm_lba configs (#146)
* Modularize dev_rlwm_lba configs

* fix imports

---------
